Moved logic of export and import files to other files, added view to import DiatopicVariation.
Moved logic of importing or exporting in the database to importers.py and exporters.py and deleted from tasks.py.

Added a view importdata to use all the importers in only one view.

Added importation view to import variation using the browser, works fine also in an asyncronic way.
