+++ linguatec_lexicon/management/commands/exportvariation.py
@@ -3,7 +3,7 @@ from django.core.management.base import BaseCommand, CommandError
 from linguatec_lexicon.models import (
     Lexicon, DiatopicVariation)
 
-from linguatec_lexicon.tasks import write_to_csv_file_export_variation
+from linguatec_lexicon.exporters import write_to_csv_file_variation
 
 import os.path
 
@@ -55,4 +55,4 @@ class Command(BaseCommand):
         if os.path.isfile(self.output_file):
             raise CommandError('Error: A csv with that name already exists: ' + self.output_file)
 
-        write_to_csv_file_export_variation.now(self.lexicon.pk, self.variation.pk, self.output_file)
+        write_to_csv_file_variation(self.lexicon.pk, self.variation.pk, self.output_file)
