+++ linguatec_lexicon/templates/linguatec_lexicon/importdata.html
@@ -1,0 +1,53 @@
+{% extends 'rest_framework/base.html' %}
+{% load i18n static %}
+
+{% block title %}{{ title }}{% endblock %}
+
+{% block branding %}
+<a class="navbar-brand" rel="nofollow" href="/">{{ title }}</a>
+{% endblock %}
+
+{% block content %}
+<h1>{{ title_variation }}</h1>
+<div class="row">
+    <div class="col-12">
+        <div>
+            Selecciona el fichero XLSX para importar las variaciones diatopicas.
+        </div>
+        <div class="data-validator-form">
+            <form action="" enctype="multipart/form-data" method="post">
+                {% csrf_token %}
+                {{ form.as_p }}
+                <label for="lexicon_code">Codigo lexicon</label><input type="text" name="lexicon_code" /></br>
+                <label for="variation_name">Nombre Variación</label><input type="text" name="variation_name" /></br>
+
+                <input name="type_export" type="hidden" value="variation">
+                <input id="Importar" type="submit" value="Importar" />
+            </form>
+        </div>
+    </div>
+</div>
+<div class="row">
+    <h3><strong>{{ is_finished }}</strong></h3>
+</div>
+
+<h1>{{ title_gramcats }}</h1>
+<div class="row">
+    <div class="col-12">
+        <div>
+            Selecciona el fichero XLSX para importar las categorias gramaticales.
+        </div>
+        <div class="data-validator-form">
+            <form action="" enctype="multipart/form-data" method="post">
+                {% csrf_token %}
+                {{ form.as_p }}
+                <input name="type_export" type="hidden" value="gramcats">
+                <input id="Importar" type="submit" value="Importar" />
+            </form>
+        </div>
+    </div>
+</div>
+<div class="row">
+    <h3><strong>{{ is_finished }}</strong></h3>
+</div>
+{% endblock %}
