--- linguatec_lexicon/templates/linguatec_lexicon/importgramcats.html
@@ -1,29 +1,0 @@
-{% extends 'rest_framework/base.html' %}
-{% load i18n static %}
-
-{% block title %}{{ title }}{% endblock %}
-
-{% block branding %}
-<a class="navbar-brand" rel="nofollow" href="/">{{ title }}</a>
-{% endblock %}
-
-{% block content %}
-<h1>{{ title }}</h1>
-<div class="row">
-    <div class="col-12">
-        <div>
-            Selecciona el fichero XLSX para importar las categorias gramaticales.
-        </div>
-        <div class="data-validator-form">
-            <form action="" enctype="multipart/form-data" method="post">
-                {% csrf_token %}
-                {{ form.as_p }}
-                <input id="validate" type="submit" value="validate" />
-            </form>
-        </div>
-    </div>
-</div>
-<div class="row">
-    <h3><strong>{{ is_finished }}</strong></h3>
-</div>
-{% endblock %}
