+++ linguatec_lexicon/urls.py
@@ -14,5 +14,5 @@ urlpatterns = [
     path('validator/', views.DataValidatorView.as_view(), name='validator'),
     path('validator-diatopic-variation/', views.DiatopicVariationValidatorView.as_view(), name='validator-variation'),
     path('exportdata/', views.ExportData.as_view(), name='exportdata'),
-    path('importgramcats/', views.ImportGramcatsView.as_view(), name='importgramcats'),
+    path('importdata/', views.ImportData.as_view(), name='importdata'),
 ]
